Typo fix in CfgScale output type, typo fix in sampler.py

+++ util.py
@@ -70,7 +70,7 @@ class HelperNodes_CfgScale(BaseNode):
             }
         }
 
-    RETURN_TYPES = ("INT", )
+    RETURN_TYPES = ("FLOAT", )
     RETURN_NAMES = ("CFG",)
 
     CATEGORY = f"{MODULE_CATEGORY}"
